Show summary about disabled optimizations
